Include list/edit functionality within the sandbox
The pipeline is that a each user is generated two sandbox collections.
The first is a free sandbox without any PBSHM Schema validation loaded
into it, the second is a validated sandbox which included the PBSHM
Schema. Models are first created within the free sandbox and go through
minimal validation (valid json & irreducible element section declared),
once a model is ready to be validated, it will then be inserted into the
validated schema for syntax verification.