Fix core module namespace clash
This module uses the same pbshm namespace as the
pbshm core, subsequently when initialising the app
any references to the core module fail as python looks
in the pbshm folder of the module instead of the pbshm
folder in the site packages. The core modules are instead
imported as ModuleSpec's from file location and manually
included into the registered system modules. This scenario
is only required when developing and testing locally before
packaging and deploying.